apps/python/pyproject.toml
[project]
name = "clientpix-api"
version = "0.1.0"
description = "Client Pix API - Self-hosted photography client gallery"
readme = "README.md"
requires-python = ">=3.13.6"
dependencies = [
    "sqlalchemy>=2.0.45",
    "asyncpg>=0.31.0",
    "pydantic>=2.12.5",
    "alembic>=1.17.2",
    "pillow>=12.1.0",
    "python-multipart>=0.0.21",
    "aiofiles>=25.1.0",
    "uvicorn>=0.40.0",
    "fastapi[standard]>=0.128.0",
    "python-dotenv>=1.2.1",
    "greenlet>=3.3.0",
]

[dependency-groups]
dev = [
    "ruff>=0.13.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
